requirements.txt
requests
python-dotenv
pymupdf
Pillow
beautifulsoup4
unstructured
langchain
langchain-community
pdfminer.six==20221105
pi_heif
unstructured_inference
pdf2image
pytesseract
unstructured[all]
unstructured_pytesseract
llama-cloud-services
llama-index-core 
llama-index-readers-file
llama-index-llms-openai
llama-index-embeddings-openai
llama-parse
faiss-cpu
langchain-openai
